Resolve the segmentation fault error related to simple_quic_server.
When the filename in the command-line arguments of simple_quic_server is a relative path,
the fopen system call fails, leading to the failure of ProofSource initialization and
subsequently causing a segmentation fault due to a null pointer reference.

Resolve: The FilePath class adds a method to check whether the file name is an absolute path.
Simple_quic_server checks if the file is an absolute path when opening the file